Fix GTFS merge cornercases when merged file does/does not exist

+++ src/gtfs-manager/gtfsmanager.py
@@ -69,12 +69,14 @@ class GTFSManager:
         self.download_gtfs(self.current_fname, url)
         if os.path.isfile(self.stored_fname):
             if self.is_gtfs_changed():
-            	print "File has changed. Merging ..."
+            	print "Merging ..."
+		if os.path.isfile(self.merged_fname):
+		    os.remove(self.merged_fname)
             	self.merge_gtfs(self.stored_fname, self.current_fname, self.merged_fname)
             	shutil.copy2(self.current_fname, self.stored_fname)
-	    else:
-		shutil.copy2(self.current_fname, self.merged_fname)
-            return
+            else:
+                if not os.path.isfile(self.merged_fname):
+                    shutil.copy2(self.current_fname, self.merged_fname)
         else:
             shutil.copy2(self.current_fname, self.stored_fname)
             shutil.copy2(self.current_fname, self.merged_fname)
